test 6

+++ frontend/src/utils/deleteProduct.ts
@@ -4,6 +4,8 @@ const deleteProduct = async (slug: string) => {
 	);
 	const data = await findProduct.json();
 
+	if (data) return;
+
 	await fetch(`${process.env.NEXT_PUBLIC_URL}/products/${data.data[0]?.id}`, {
 		method: 'DELETE',
 		headers: {
